+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 # 1.0.0
 - Initial release
+
+# 1.1.0
+- Maven root module creation
+- Maven submodule creation
+- VSC launch configuration file creation